Ghosts can't haunt canvases (#91759)
## About The Pull Request

Ghosts are snowflaked able to interact with the canvas UI, which meant
they were able to press buttons including finalize. This fixes that with
a special observer check. Leaves a comment to explain why it is the way
it is as well in case someone in the future messes it up.

## Why It's Good For The Game

Closes https://github.com/tgstation/tgstation/issues/91750

## Changelog

:cl:
fix: Observers can no longer finalize unfinished canvas paintings.
/:cl: