Fixes CatwalkStation armory not having energy guns (#91926)
## About The Pull Request

Actually adds eguns to the gunlocker in armory, because I thought that
it was a preset and not container-like in mapping.

## Why It's Good For The Game

In making of #91922, I noticed that the gunlocker in the armory didnt
have any eguns in it, and since I was the one who made that change, I'll
be the one to fix it before it inconveniences anyone.

## Changelog

:cl:

fix: Fixes CatwalkStation's armory not having eguns in the gunlocker.

/:cl: