Fix xenobio console slime duplication (#92274)
## About The Pull Request

So it seems like it was possible for the `slime_pickup(...)` proc to be
called on a slime we were already picking up when spam-clicking,
especially on high time dilation, leading to the slime being added to
the list of stored slimes twice.
This just adds an early return in case we have already added this slime
to our list.
This fixes our issue.
## Why It's Good For The Game

Fixes #92270.
Fixes #91909.
Fixes #91686.
## Changelog
:cl:
fix: Xenobio console can no longer pick up the same slime multiple
times.
/:cl: