Add custom sound action on command report (#91985)
## About The Pull Request

When I was administering one downstream there were often situations when
instead of the standard announcement sound it was necessary to play
something of my own. So I got my hands on it. Together with standard
announcement sounds there is a possibility to play a custom sound.


https://github.com/user-attachments/assets/ba17f4d1-56f6-4022-b049-6ad6dab8481f


## Why It's Good For The Game

Admins will ~~get a soundpad~~ have the ability to play their own
announcement sounds to match the tone of the situation.

## Changelog
:cl:
admin: Added the ability to play your own announce sounds.
/:cl: