Fixes cyborgs taking burn damage from lasers, and brute from EMP (#92337)
## About The Pull Request
Used to be, cyborgs would take brute damage from ranged burn attacks,
like lasers and emitters and whatnot. The intention was that these dealt
`chassis` damage, and burn was being reused as `wire` damage. So EMPs
damaged internal wiring, lasers hit the outer shell.

This was changed in #79024, due to a refractoring of the proc, but it's
not listed in the changelog, so I believe it was an oversight.

On an unrelated note, while testing this and the fix I made, I found
that EMP damage was incorrectly being dealt to silicons as
brute/`chassis` damage, due to how the damage was called. So that's also
been fixed.