The Chain of Command can now give a speed boost by whipping other players (#92387)
## About The Pull Request

The Chain of Command can now hasten other mobs when right-click
attacked.



https://github.com/user-attachments/assets/b6c2185e-0edb-45d5-a015-3c7e0a684ffc



It's not a very ample speed boost, but it's enough to make it worth
taking into battle if you have friendlies to buff. Lasts 7 seconds.
Refreshes on hit. You cannot whip yourself.

This only applies to the Chain of Command, not whips nor tailwhips.
There's only one of these bad boys on the station. Use it well.
## Why It's Good For The Game

The chain of command is nice as a ceremonial weapon, but it doesn't
actually see any use in combat. Ever.

Instead of just making it do more damage (boring), I wanted to make it a
support tool. The captain (its owner) should usually be in the backlines
of a fight, after all.

Support tools gives less combat-focused players (or pacifists) a way
contribute meaningfully to a team fight (nukies, revs). Not equipped to
fight the blob with the rest of the team? You can use this to buff your
allies without putting yourself within the blob's reach.

<details>
<summary>The EVIL secret reason I think it's good for the
game...</summary>
It gives people some mechanical incentive to steal the chain of command
for once >:)
</details>


## Changelog
:cl: Rhials
balance: The Chain of Command will give a speed bonus to other players
when right-clicking them.
/:cl: