Adds warning that you can't absorb people you've used DNA extract on, and makes some descriptions more consistent (#92397)
## About The Pull Request

Adds a warning in the DNA Extract Sting helptext that you won't be able
to absorb bodies you've stung with it.
Also changes several "you"s and "your"s in the descriptions into the
proper "we"s and such.

## Why It's Good For The Game

closes #92373 

Most of them already use "we"s "us"s and "our"s, and its oddly
inconsistent as is