Fixes CRAB-17 sometimes not draining funds (#92335)
## About The Pull Request

Fixes #92230

Fixes an issue where the CRAB-17 would sometimes fail to drain funds due
to being unable to send money to an account.

In `dump()`:

```
var/datum/bank_account/account = bogdanoff?.get_bank_account()
if (account) // get_bank_account() may return FALSE
	account.transfer_money(B, amount, "?VIVA¿: !LA CRABBE¡")
	B.bank_card_talk("You have lost [percentage_lost * 100]% of your funds! A spacecoin credit deposit machine is located at: [get_area(src)].")
```

If during any call of `dump()` our Bogdanoff takes off their ID, never
had one to begin with, or otherwise makes their bank account not visible
to `get_bank_account()` the `if(account)` check fails and money is never
drained.

Instead this PR adds an internal bank account to the CRAB-17, and any
time `dump()` fails to find a destination account the money is stored in
the CRAB-17's account. When destroyed any funds stored internally are
scattered across the floor in a pile of smaller holochips.

Incidentally adds documentation to the relevant code.

## Why It's Good For The Game

Players expect the CRAB-17 to drain funds even if they don't have an
account to send the money to. It's still preferable to have one so they
get the money, but smearing it across the floor still forces the crew to
fight over who gets it, and is more logical than sending it to the void.